Wallet enhancements, remove client-side signing.
