+++ packages/client-core/src/user/components/UserMenu/menus/ProfileMenu.tsx
@@ -6,7 +6,7 @@ import { useTranslation } from 'react-i18next'
 import { useLocation } from 'react-router-dom'
 
 import { validateEmail, validatePhoneNumber } from '@xrengine/common/src/config'
-import { requestVcForEvent } from '@xrengine/common/src/credentials/credentials'
+import { requestVcForEvent, vpRequestQuery } from '@xrengine/common/src/credentials/credentials'
 import multiLogger from '@xrengine/common/src/logger'
 import { AudioEffectPlayer } from '@xrengine/engine/src/audio/systems/MediaSystem'
 import { WorldState } from '@xrengine/engine/src/networking/interfaces/WorldState'
@@ -257,26 +257,6 @@ const ProfileMenu = ({
    * Example function, requests a Verifiable Credential from the user's wallet.
    */
   async function handleRequestCredentialClick() {
-    const vpRequestQuery: any = {
-      web: {
-        VerifiablePresentation: {
-          query: [
-            {
-              type: 'QueryByExample',
-              credentialQuery: [
-                {
-                  example: {
-                    '@context': ['https://www.w3.org/2018/credentials/v1', 'https://w3id.org/xr/v1'],
-                    type: 'VerifiableCredential'
-                  }
-                }
-              ]
-            }
-          ]
-        }
-      }
-    }
-
     const result = await navigator.credentials.get(vpRequestQuery)
 
     console.log('VC Request query result:', result)
