+++ packages/client-core/src/user/services/AuthService.ts
@@ -450,7 +450,7 @@ export const AuthService = {
         identityProvider: {
           id: 0,
           token: '',
-          type: 'chapiWallet',
+          type: 'didWallet',
           isVerified: true,
           userId: walletUser.id
         }
