+++ packages/common/src/dbmodels/UserInterface.ts
@@ -3,4 +3,5 @@ export interface UserInterface {
   name: string
   inviteCode: string
   isGuest: boolean
+  did: string
 }
