+++ packages/common/src/dbmodels/VerificationEvent.ts
@@ -2,4 +2,5 @@ export interface VerificationEventInterface {
   id: String
   userId: String
   expiresAt: Date
+  log: String // verification result log (fine-grained results, like "expired" / "revoked"
 }
