+++ packages/server-core/src/credential/verification-event/verification-event.model.ts
@@ -27,6 +27,10 @@ export default (app: Application) => {
         type: DataTypes.DATE,
         defaultValue: Date.now(),
         allowNull: false
+      },
+      log: {
+        type: DataTypes.STRING,
+        allowNull: true
       }
     },
     {
